Move service collection extensions
