+++ src/Altinn.Notifications.Email.Core/Altinn.Notifications.Email.Core.csproj
@@ -8,6 +8,12 @@
 		<ProjectGuid>{8823A1D7-BFA1-40EC-B9D5-C1032FD7133F}</ProjectGuid>
 	</PropertyGroup>
 
+	<ItemGroup>
+	  <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="7.0.0" />
+	  <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="7.0.0" />
+	  <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.1" />
+	</ItemGroup>
+
 	<ItemGroup Condition="'$(Configuration)'=='Debug'">
 		<PackageReference Include="StyleCop.Analyzers" Version="1.2.0-beta.507">
 			<PrivateAssets>all</PrivateAssets>
@@ -16,10 +22,6 @@
 		<AdditionalFiles Include="..\..\stylecop.json">
 			<Link>stylecop.json</Link>
 		</AdditionalFiles>
-	</ItemGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.1" />
 	</ItemGroup>
 
 	<PropertyGroup Condition="'$(Configuration)'=='Debug'">
