+++ src/Altinn.Notifications.Email.Integrations/Altinn.Notifications.Email.Integrations.csproj
@@ -10,8 +10,13 @@
 
 	<ItemGroup>
 		<PackageReference Include="Confluent.Kafka" Version="2.2.0" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="7.0.4" />
 		<PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="7.0.1" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\Altinn.Notifications.Email.Core\Altinn.Notifications.Email.Core.csproj" />
 	</ItemGroup>
 
 	<ItemGroup Condition="'$(Configuration)'=='Debug'">
@@ -22,10 +27,6 @@
 		<AdditionalFiles Include="..\..\stylecop.json">
 			<Link>stylecop.json</Link>
 		</AdditionalFiles>
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Altinn.Notifications.Email.Core\Altinn.Notifications.Email.Core.csproj" />
 	</ItemGroup>
 
 	<PropertyGroup Condition="'$(Configuration)'=='Debug'">
